Reach 23.53% matching C coverage with parallel decompilation [skip ci]

+++ include/unk_02031B0C.h
@@ -18,7 +18,7 @@ void sub_02031CCC(SaveApricornBox *, int);
 BOOL sub_02031CEC(SaveApricornBox *apricornBox, u16 a1, UnkStruct_02031CEC *a2);
 int sub_02031D80(UnkStruct_02031CEC *a0);
 void sub_02032058(SaveApricornBox *saveApricornBox, u32 steps);
-String *sub_020322AC(SaveApricornBox *apricornBox, u16 a1, u32 a2);
+String *sub_020322AC(SaveApricornBox *apricornBox, int a1, u32 a2);
 String *sub_02032308(SaveApricornBox *apricornBox, u16 a1, u32 a2);
 int ApricornBox_CountApricorn(SaveApricornBox *apricornBox, int a1);
 int ApricornBox_GiveApricorn(SaveApricornBox *apricornBox, int a1, u8 a2);
